# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI Operational POS API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-UNLICENSED
  description: Phase 0 foundation contract with the backend voice transcription slice. Product, auth, sales, inventory, and intent handlers are not declared until implemented.
servers:
  - url: http://127.0.0.1:8080
security: []
paths:
...
  /api/v1/assistant/transcriptions:
    post:
      operationId: createAssistantTranscription
      summary: Transcribe an uploaded audio clip
      description: Transcribes audio in memory through the configured speech-to-text provider. Audio is not persisted.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TranscriptionRequest'
      responses:
        '200':
          description: Audio was transcribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessTranscription'
        '400':
          description: The multipart request or duration field is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The audio exceeds the request size or duration limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: The audio MIME type or extension is unsupported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: The transcription provider failed or returned an invalid response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/uploads:
    post:
      operationId: uploadImage
      summary: Upload a product image
      description: Uploads a WebP, PNG, or JPEG product image and stores it on disk.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Product image file (webp, png, jpeg).
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessUpload'
        '400':
          description: Missing file or file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unsupported file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/products:
    get:
      operationId: listProducts
      summary: List all active products
      description: Returns all active products in the catalog ordered alphabetically by name.
      responses:
        '200':
          description: List of active products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessProductList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createProduct
      summary: Create a new product in the catalog
      description: Validates and creates a new product with unique SKU and non-negative pricing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessProductCreated'
        '400':
          description: Validation error in request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Product with this SKU already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Voice transcription is disabled or not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/inventory/locations:
    get:
      operationId: listInventoryLocations
      summary: List all active inventory locations
      description: Returns active storage locations (e.g., Toko Utama, Gudang).
      responses:
        '200':
          description: List of inventory locations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessLocationList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/inventory/stocks:
    get:
      operationId: listStocks
      summary: List current stock balances
      description: Returns stock balances per product and inventory location.
      parameters:
        - name: location_id
          in: query
          required: false
          schema:
            type: string
          description: Optional filter by inventory location UUID.
      responses:
        '200':
          description: List of stock balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessStockList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/inventory/movements:
    post:
      operationId: createStockMovement
      summary: Record an inventory stock movement
      description: Appends a stock movement entry and updates the product stock balance atomically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordMovementRequest'
      responses:
        '201':
          description: Stock movement recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMovementRecorded'
        '400':
          description: Invalid movement data or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders:
    get:
      operationId: listOrders
      summary: List past orders
      description: Returns a list of orders optionally filtered by location.
      parameters:
        - name: location_id
          in: query
          required: false
          schema:
            type: string
          description: Optional filter by inventory location UUID.
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOrderList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createOrder
      summary: Create and complete a POS order
      description: Persists order and items, deducts inventory stock atomically, and appends sale stock movement.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created and processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOrderDetail'
        '409':
          description: Insufficient inventory stock for order items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error or insufficient payment amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{id}:
    get:
      operationId: getOrderById
      summary: Get order details and items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order UUID or human-readable order number.
      responses:
        '200':
          description: Detailed order with items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOrderDetail'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      operationId: loginWithPassword
      summary: Login with email and password
      description: Verifies staff credentials with bcrypt and issues an opaque session token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session with user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuth'
        '400':
          description: Email or password missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/pin:
    post:
      operationId: loginWithPin
      summary: Fast cashier terminal login with role and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinLoginRequest'
      responses:
        '200':
          description: Authenticated session with user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuth'
        '401':
          description: Invalid role or PIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Role or PIN payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      operationId: getSessionUser
      summary: Resolve the current session token to a user
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer opaque session token.
      responses:
        '200':
          description: Session user with role and permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthUser'
        '401':
          description: Missing, revoked, or expired session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      operationId: revokeSession
      summary: Revoke the current session token
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer opaque session token.
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessLogout'
        '401':
          description: Missing or unknown session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/categories:
    get:
      operationId: listCategories
      summary: List active product categories for the tenant
      responses:
        '200':
          description: Tenant category list ordered by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCategoryList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createCategory
      summary: Create a product category
      description: Category names are unique per tenant (case-insensitive).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCategoryCreated'
        '400':
          description: Category name missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Category name already exists for the tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/customers:
    get:
      operationId: listCustomers
      summary: List active customers for the tenant
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Filter by name, phone, or email (case-insensitive).
      responses:
        '200':
          description: Tenant customer list ordered by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomerList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createCustomer
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '400':
          description: Customer name missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/customers/{id}:
    get:
      operationId: getCustomerById
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateCustomer
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '400':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/pets:
    get:
      operationId: listPets
      summary: List active pets for the tenant
      parameters:
        - name: customer_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by owner customer ID.
      responses:
        '200':
          description: Tenant pet list ordered by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPetList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createPet
      summary: Register a new pet for a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPetRequest'
      responses:
        '201':
          description: Pet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPet'
        '422':
          description: Payload invalid or owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/bookings/{id}:
    get:
      operationId: getBookingById
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessBooking'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    TranscriptionRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Audio file with a supported extension and matching MIME type.
        duration_seconds:
          type: number
          format: double
          minimum: 0
          description: Optional client-observed duration used for an early duration-limit check.
    SuccessTranscription:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [text, provider, model]
          properties:
            text:
              type: string
            provider:
              type: string
              const: groq
            model:
              type: string
              description: Configured speech-to-text model identifier.
        request_id:
          type: string
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        request_id:
          type: string
...
    SuccessUpload:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [url, filename, size, content_type]
          properties:
            url:
              type: string
            filename:
              type: string
            size:
              type: integer
            content_type:
              type: string
        request_id:
          type: string
...